requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
imbalanced-learn
joblib
pymysql
sqlalchemy
streamlit
shap
plotly
python-dotenv